+++ src/revars.ts
@@ -2,20 +2,64 @@
   regex list
 */
 
-// name=value in directives - $1:name, $2:function, $3:value (including any comment)
-export const VARPAIR = /^\s*(_[0-9A-Z][_0-9A-Z]*)\s*(?:\(\s*([^)]+?)\s*\))?\s*=?(.*)/
+/**
+ * Stringified regex to compose another regex that supports single and double
+ * quoted strings.
+ * This is some complex regex, but is the most efficient implementation in JS.
+ */
+const STR_BASE = /"[^"\n\r\\]*(?:\\(?:\r\n?|[\S\s])[^"\n\r\\]*)*"/.source
 
-// to verify valid varnames and for #unset
+/**
+ * Test for valid jscc varnames.
+ */
 export const VARNAME = /^_[0-9A-Z][_0-9A-Z]*$/
 
-// prefixing varnames inside expression with `this.` or `global.`
-export const EVLVARS = /(^|[^$\w.])(_[0-9A-Z][_0-9A-Z]*)\b(?=[^$\w]|$)/g
+/**
+ * Matches jscc var assignments in the format "_VAR=expr" of jscc directives.
+ *
+ * - $1: varname
+ * - $2: value (well... rest of the chars, including any comment)
+ */
+export const ASSIGNMENT = /^\s*(_[0-9A-Z][_0-9A-Z]*)\s*=?(.*)/
 
-// replace varnames inside the code from $_VAR.prop to value
-export const REPVARS = /(?:(\$_[0-9A-Z][_0-9A-Z]*)(?:\(\s*([^)]+?)\s*\))?([.\w]+)?)(?=[\W]|$)/g
+/**
+ * Regex to search varnames (format "_VAR") inside jscc expressions.
+ *
+ * __NOTE:__
+ *
+ * This function allows the varname to be followed by any char that is not
+ * not part of another jscc varname or JS variable name. Because this, it
+ * supports properties (ex: `_VAR.prop` or `_VAR['prop']`) to be evaluated
+ * by the `evalExpr` function.
+ *
+ * - $1: char before varname, not in the set [$\w.]
+ * - $2: varname
+ */
+export const JSCC_VARS = /(^|[^$\w.])(_[0-9A-Z][_0-9A-Z]*)(?=[^$0-9a-z]|$)/g
 
-// for nested objects inside REPVARS
-export const PROPVARS = /\.(\w+)/g
+/**
+ * Regex to search varnames in the format "$_VAR" followed by zero or more
+ * properties with dot notation _inside_ the code already processed.
+ *
+ * __NOTE:__
+ *
+ * This regex allows var concatenation like in `$_VAR1$_VAR2`
+ *
+ * Code supporting macro replacement was removed in v1.0. It still needed more
+ * work to skip nested braces, strings, regexes, ES6 TLS... and I'm not sure if
+ * this feature is necessary, cannot find a valid use case to keep it.
+ *
+ * - $1: var name
+ * - $2: optional expression
+ *
+ * TODO: More testing on multiple properties and concatenation.
+ */
+export const VARS_TO_REPL = /(?:(\$_[0-9A-Z][_0-9A-Z]*)((?:\.\w+)+)*)(?=\W|$)/g
 
-// matches single and double quoted strings, take care about embedded eols
-export const STRINGS = /"[^"\n\r\\]*(?:\\(?:\r\n?|[\S\s])[^"\n\r\\]*)*"|'[^'\n\r\\]*(?:\\(?:\r\n?|[\S\s])[^'\n\r\\]*)*'/g
+/**
+ * Matches single and double quoted strings taking care of embedded (escaped)
+ * quotes and EOLs of multiline strings.
+ *
+ * It has no captures.
+ */
+export const STRINGS = RegExp(STR_BASE + '|' + STR_BASE.replace(/"/g, '"'), 'g')